requirements.txt
backports.entry-points-selectable==1.1.1
click==8.0.3
colorama==0.4.4
distlib==0.3.4
dnspython==2.1.0
email-validator==1.1.2
filelock==3.4.0
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
Flask-Validator==1.4.2
greenlet==1.1.2
idna==3.3
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0
platformdirs==2.4.0
protobuf==3.19.1
py-moneyed==0.8.0
pycountry==20.7.3
pytz==2020.1
schwifty==2020.9.0
six==1.16.0
SQLAlchemy==1.3.24
virtualenv==20.10.0
Werkzeug==2.0.2
yapf==0.31.0
